requirements.txt
python-telegram-bot==21.3
telethon==1.36.0
cryptography==42.0.8
psutil==5.9.8
